Change CRUD for threads to support user validation

+++ controllers/thread.go
@@ -30,11 +30,15 @@ func GetThreads(context *gin.Context) {
 func CreateThread(context *gin.Context) {
 	var newThread models.Thread
 
-	// TODO: validate author
-	if err := context.BindJSON(&newThread); err != nil {
+	user := context.MustGet("user").(models.User)
+
+	if err := context.ShouldBindJSON(&newThread); err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"error": "Failed to parse request body"})
 		return
 	}
+
+	// Tag the thread with the author
+	newThread.Author = user
 
 	// Add new thread to the database
 	result := initializers.DB.Create(&newThread)
@@ -47,7 +51,7 @@ func CreateThread(context *gin.Context) {
 }
 
 func GetThread(context *gin.Context) {
-	id, err := strconv.Atoi(context.Param("id"))
+	id, err := strconv.Atoi(context.Param("threadID"))
 
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID format"}) //TODO: abstract out invalid integer error message
@@ -65,7 +69,7 @@ func GetThread(context *gin.Context) {
 }
 
 func UpdateThread(context *gin.Context) {
-	id, err := strconv.Atoi(context.Param("id"))
+	id, err := strconv.Atoi(context.Param("threadID"))
 
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID format"}) //TODO: abstract out invalid integer error message
@@ -79,6 +83,15 @@ func UpdateThread(context *gin.Context) {
 		return
 	}
 
+	// check if user is author of thread
+	user := context.MustGet("user").(models.User)
+
+	if thread.AuthorID != user.ID {
+		context.IndentedJSON(http.StatusUnauthorized, gin.H{"message": "Unauthorized. You are not the author of this thread"})
+		return
+	}
+
+	// Update thread
 	var updatedThread models.Thread
 
 	if err := context.BindJSON(&updatedThread); err != nil {
@@ -95,7 +108,7 @@ func UpdateThread(context *gin.Context) {
 
 func DeleteThread(context *gin.Context) {
 	// TODO: ensure only author can delete thread
-	id, err := strconv.Atoi(context.Param("id"))
+	id, err := strconv.Atoi(context.Param("threadID"))
 
 	if err != nil {
 		context.JSON(http.StatusBadRequest, gin.H{"error": "Invalid ID format"}) //TODO: abstract out invalid integer error message
@@ -109,6 +122,18 @@ func DeleteThread(context *gin.Context) {
 		return
 	}
 
+	// check if user is author of thread
+	user := context.MustGet("user").(models.User)
+
+	if thread.AuthorID != user.ID {
+		context.IndentedJSON(http.StatusUnauthorized, gin.H{"message": "Unauthorized. You are not the author of this thread"})
+		return
+	}
+
+	// Delete thread from user
+	initializers.DB.Model(&user).Association("Threads").Delete(&thread)
+
+	// Delete thread
 	initializers.DB.Delete(&thread, id)
 
 	context.IndentedJSON(http.StatusOK, thread)
